src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/content-pseudo-element-overridden-expected.txt

CONSOLE ERROR: line 33: Uncaught TypeError: host.createShadowRoot is not a function
content

Test for crbug.com/274059. Should be able to override ::content styles in shadow root style sheet from the document.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
